requirements.txt
absl-py==2.1.0
astunparse==1.6.3
brewer2mpl==1.4.1
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
cloudpickle==3.0.0
coloredlogs==15.0.1
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
dm-tree==0.1.8
filelock==3.14.0
flatbuffers==24.3.25
fonttools==4.53.0
fsspec==2024.6.0
ftfy==6.2.0
gast==0.5.4
ggplot==0.11.5
google-auth==2.30.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.64.1
h5py==3.11.0
huggingface-hub==0.23.4
humanfriendly==10.0
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
keras==2.15.0
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
ml-dtypes==0.3.2
moepy==1.1.4
mpmath==1.3.0
namex==0.0.8
networkx==3.3
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnx==1.16.1
onnx-tf==1.10.0
onnxruntime==1.18.0
opt-einsum==3.3.0
optree==0.11.0
packaging==24.1
pandas==2.2.2
patsy==0.5.6
pillow==10.2.0
progress==1.6
protobuf==3.20.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
Pygments==2.18.0
pyparsing==3.1.2
PyQt6==6.7.0
PyQt6-Qt6==6.7.1
PyQt6-sip==13.6.0
python-dateutil==2.9.0.post0
pytorch-triton-rocm==2.3.1
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.14.0
seaborn==0.13.2
sentence-transformers==3.0.1
six==1.16.0
sortedcontainers==2.4.0
statsmodels==0.14.2
sympy==1.12.1
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow-addons==0.23.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.0
tensorflow-probability==0.23.0
termcolor==2.4.0
tf2onnx==1.16.1
tf_keras==2.15.1
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.1+rocm6.0
torchaudio==2.3.1+rocm6.0
torcheval==0.0.7
torchvision==0.18.1+rocm6.0
tqdm==4.66.4
transformers==4.42.3
triton==2.3.1
typeguard==2.13.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.14.1
